+++ responder/routes.py
@@ -1,6 +1,5 @@
 import re
-from graphql import GraphQLSchema
-from parse import parse, search
+from parse import parse
 
 
 def memoize(f):
@@ -14,6 +13,8 @@ def memoize(f):
 
 
 class Route:
+    _param_pattern = re.compile(r"{([^{}]*)}")
+
     def __init__(self, route, endpoint):
         self.route = route
         self.endpoint = endpoint
@@ -36,7 +37,7 @@ class Route:
 
     @property
     def has_parameters(self):
-        return all([("{" in self.route), ("}" in self.route)])
+        return bool(self._param_pattern.search(self.route))
 
     @memoize
     def does_match(self, s):
@@ -59,9 +60,10 @@ class Route:
         return url
 
     def _weight(self):
-        params_count = -len(set(re.findall(r'{([a-zA-Z]\w*)}', self.route)))
+        params = set(self._param_pattern.findall(self.route))
+        params_count = -len(params) or 0
         return params_count != 0, params_count
 
     @property
     def is_graphql(self):
-        return isinstance(self.endpoint, GraphQLSchema)
+        return hasattr(self.endpoint, "get_graphql_type")